feat(Restaurar clave)(Restaurar clave): :construction:

+++ OdontoSoftBack/src/main/java/org/enterprise/odontosoft/model/Dao/UsuarioDao.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UsuarioDao extends CrudRepository<Usuario, Long> {
+public interface UsuarioDao extends CrudRepository<Usuario, Integer> {
 
   @Query(value = "select u from Usuario u where u.codigo = ?1 and u.habilitado = true")
   public Usuario findByCodigo(String codigo);
